+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Unreleased
+
+### 3.6.0 2018-09-11
   - Added support for QCAS form types.
   - Calculate total values for QCAS to be sent downstream.
   - Replace all negative numbers with a string of 11 characters containing only 9's in pck file.